// usage/generated/common/inputs.ts
export class StringFieldUpdateOperationsInput {
  @NestJsGraphQL.Field(() => String, { nullable: true })
  set?: string | undefined;
}

@NestJsGraphQL.InputType('NullableStringFieldUpdateOperationsInput', { isAbstract: true })
export class NullableStringFieldUpdateOperationsInput {
  @NestJsGraphQL.Field(() => String, { nullable: true })
  set?: string | undefined;
}

@NestJsGraphQL.InputType('DateTimeFieldUpdateOperationsInput', { isAbstract: true })
export class DateTimeFieldUpdateOperationsInput {
  @NestJsGraphQL.Field(() => Date, { nullable: true })
  set?: Date | undefined;
}